Allow migrations to run with empty database
This migration could not be run with a fresh database because it was
relying on the “Curiosity” rover to exist in the database.

It was also relying on the existence of a `Rover` class, which has now
been defined at the top of the migration to ensure that even if this
class is renamed/moved etc., this migration will still work.